+++ src/Http/Resources/SalesQuotationResource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ChrisLorando\LaravelAccurate\Http\Resources;
+
+class SalesQuotationResource extends Resource
+{
+    protected string $resourceName = 'sales-quotation';
+}